pyproject.toml
[project]
name = "fastmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "fastapi-mcp>=0.4.0",
    "openai>=2.7.1",
    "pydantic>=2.12.4",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
